+++ enterprise/api/v1/serializers.py
@@ -547,6 +547,7 @@ class EnterpriseCustomerCourseEnrollmentsSerializer(serializers.Serializer):
     tpa_user_id = serializers.CharField(required=False, write_only=True)
     user_email = serializers.EmailField(required=False, write_only=True)
     course_run_id = serializers.CharField(required=True, write_only=True)
+    cohort = serializers.CharField(required=False, write_only=True)
     course_mode = serializers.ChoiceField(
         choices=(
             ('audit', 'audit'),
@@ -558,6 +559,7 @@ class EnterpriseCustomerCourseEnrollmentsSerializer(serializers.Serializer):
     )
     email_students = serializers.BooleanField(default=False, required=False, write_only=True)
     detail = serializers.CharField(read_only=True)
+    is_active = serializers.BooleanField(required=False, default=True, write_only=True)
 
     def create(self, validated_data):
         """
@@ -569,25 +571,35 @@ class EnterpriseCustomerCourseEnrollmentsSerializer(serializers.Serializer):
         user_email = validated_data.get('user_email')
         course_run_id = validated_data.get('course_run_id')
         course_mode = validated_data.get('course_mode')
+        cohort = validated_data.get('cohort')
         email_students = validated_data.get('email_students')
+        is_active = validated_data.get('is_active')
 
         enterprise_customer_user = lms_user or tpa_user or user_email
 
         if isinstance(enterprise_customer_user, models.EnterpriseCustomerUser):
             validated_data['enterprise_customer_user'] = enterprise_customer_user
             try:
-                enterprise_customer_user.enroll(course_run_id, course_mode)
+                if is_active:
+                    enterprise_customer_user.enroll(course_run_id, course_mode, cohort=cohort)
+                else:
+                    enterprise_customer_user.unenroll(course_run_id)
             except (utils.CourseEnrollmentDowngradeError, HttpClientError) as exc:
                 validated_data['detail'] = str(exc)
                 return validated_data
 
-            track_enrollment('enterprise-customer-enrollment-api', enterprise_customer_user.user_id, course_run_id)
+            if is_active:
+                track_enrollment('enterprise-customer-enrollment-api', enterprise_customer_user.user_id, course_run_id)
         else:
-            enterprise_customer_user = enterprise_customer.enroll_user_pending_registration(
-                user_email,
-                course_mode,
-                course_run_id
-            )
+            if is_active:
+                enterprise_customer_user = enterprise_customer.enroll_user_pending_registration(
+                    user_email,
+                    course_mode,
+                    course_run_id,
+                    cohort=cohort
+                )
+            else:
+                enterprise_customer.clear_pending_registration(user_email, course_run_id)
 
         if email_students:
             enterprise_customer.notify_enrolled_learners(